InternalMath.PowMod returns uint

+++ Source/ac-library-csharp/Math/Internal/InternalMath.cs
@@ -163,7 +163,7 @@ namespace AtCoder.Internal
         /// <inheritdoc cref="MathLib.PowMod(long, long, int)" />
         /// </summary>
         [MethodImpl(256)]
-        public static long PowMod(long x, long n, int m)
+        public static uint PowMod(long x, long n, int m)
         {
             Contract.Assert(0 <= n && 1 <= m, reason: $"0 <= {nameof(n)} && 1 <= {nameof(m)}");
             if (m == 1) return 0;
